rename multisig proof to multisig
