+++ contracts/src/index.ts
@@ -6,4 +6,4 @@ export * from "./utils/Faucet.js";
 export * from "./pool/PoolFactory.js";
 export * from "./pool/PoolTokenHolder.js";
 export * from "./pool/Pool.js";
-export * from "./pool/MultisigProof.js";
+export * from "./pool/Multisig.js";
